+++ db/migrate/20180806154052_add_coda_reference_to_suppliers.rb
@@ -1,5 +1,0 @@
-class AddCodaReferenceToSuppliers < ActiveRecord::Migration[5.2]
-  def change
-    add_column :suppliers, :coda_reference, :string, limit: 7
-  end
-end