+++ spec/models/offboard/remove_suppliers_from_lots/row_spec.rb
@@ -11,13 +11,11 @@ RSpec.describe Offboard::RemoveSuppliersFromLots::Row do
 
     let(:supplier_name) { 'Big Dog Limited' }
     let(:salesforce_id) { 'salesforce123' }
-    let(:coda_reference) { 'C012345' }
 
     let!(:supplier) do
       supplier = create(:supplier,
                         name: supplier_name,
-                        salesforce_id: salesforce_id,
-                        coda_reference: coda_reference)
+                        salesforce_id: salesforce_id)
 
       agreement = supplier.agreements.create!(framework: framework)
 
@@ -32,7 +30,6 @@ RSpec.describe Offboard::RemoveSuppliersFromLots::Row do
         lot_number: lot_number,
         supplier_name: supplier_name,
         salesforce_id: salesforce_id,
-        coda_reference: coda_reference,
       )
     end